Fix emulated mask()
Problem: 0/1 values are shifted inside a 32Bit integer value up to 63 positions to the left. Therefore, all bits at positions > 32 in the final uint64_t return value will be set to 0.

Fix: Explicit cast to (uint64_t) before shifting.

+++ include/xsimd/arch/xsimd_emulated.hpp
@@ -508,7 +508,7 @@ namespace xsimd
             constexpr size_t size = batch<T, A>::size;
             uint64_t res = 0;
             for (size_t i = 0; i < size; ++i)
-                res |= (self.data[i] ? 1u : 0u) << i;
+                res |= (uint64_t)(self.data[i] ? 1u : 0u) << i;
             return res;
         }
